requirements.txt
gradio==4.37.1
requests==2.23.0
Pillow==10.4.0